Add files via upload
The Factory Design Pattern is a creational pattern that provides an interface for creating objects without exposing the creation logic to the client. Instead of directly instantiating classes with the new keyword, the factory method returns objects based on input or conditions. This promotes loose coupling, flexibility, and scalability, as the client code depends only on the factory, not on concrete classes.